+++ resources/views/users/followers.blade.php
@@ -9,15 +9,36 @@
     <p class='name'>{{ $user->name }}</p></br>
     <p class="info">{{ $user->height }}/{{ $user->gender }}/{{ $user->age }}</p>
     </div>
-    {!! Form::open(['route' =>['users.edit',$user->id], 'method' => 'get']) !!}
-        {!! Form::submit('プロフィール変更',['class' => "btn btn-default btn-md"]) !!}
-    {!! Form::close() !!}
+    @if (Auth::check()) 
+        @if( Auth::id() == $user->id)
+            {!! Form::open(['route' =>['users.edit',$user->id], 'method' => 'get']) !!}
+                {!! Form::submit('プロフィール変更',['class' => "btn btn-default btn-md"]) !!}
+            {!! Form::close() !!}
+        @else
+            @if (Auth::id() != $user->id)
+                @if (Auth::user()->following($user->id))
+                    {!! Form::open(['route' => ['user.unfollow', $user->id], 'method' => 'delete']) !!}
+                        {!! Form::submit('フォロー外す', ['class' => "btn btn-default btn-md"]) !!}
+                    {!! Form::close() !!}
+                @else
+                    {!! Form::open(['route' => ['user.follow', $user->id]]) !!}
+                        {!! Form::submit('フォローする', ['class' => 'btn btn-default btn-md']) !!}
+                    {!! Form::close() !!}
+                @endif
+            @endif
+        @endif
+    @else
+        {!! Form::open(['route' => 'login', 'method' => 'get']) !!}
+            {!! Form::submit('フォローする', ['class' => "btn btn-default btn-md"]) !!}
+        {!! Form::close() !!}
+    @endif
+        
     <div class='nav1'>
     <ul class="nav nav-tabs nav-justified ">
-        <li class="nav-item">{!! link_to_route('user.mypage','投稿'.count($user->fashions),['id' => $user->id],['class'=> "nav-link"]) !!}</li>
-        <li class="nav-item">{!! link_to_route('user.followings','フォロー'.count($user->followings),['id' => $user->id],['class'=> "nav-link "]) !!}</li>
+        <li class="nav-item">{!! link_to_route('users.show','投稿'.count($user->fashions),['id' => $user->id],['class'=> "nav-link" ]) !!}</li>
+        <li class="nav-item">{!! link_to_route('user.followings','フォロー'.count($user->followings),['id' => $user->id],['class'=> "nav-link"]) !!}</li>
         <li class="nav-item">{!! link_to_route('user.followers','フォロワー'.count($user->followers),['id' => $user->id],['class'=> "nav-link active"]) !!}</li>
-        <li class="nav-item">{!! link_to_route('user.favorites','お気に入り'.count($user->favorites),['id' => $user->id],['class'=> "nav-link"]) !!}</li>
+        <li class="nav-item">{!! link_to_route('user.favorites','お気に入り'.count($user->favorites),['id' => $user->id],['class'=> "nav-link "]) !!}</li>
     </ul>
     
     @if (count($followers) > 0)
@@ -28,6 +49,7 @@
                         <a href="{{ action('UsersController@show', $follower->id) }}"><img class="icon" src="/storage/image/{{ $follower->user_photo }}"></a>
                         <p>{{ $follower->name }}</p>
                         <dd>
+                    @if (Auth::check())
                         @if (Auth::id() != $follower->id)
                             @if (Auth::user()->following($follower->id))
                                 {!! Form::open(['route' => ['user.unfollow', $follower->id], 'method' => 'delete']) !!}
@@ -38,7 +60,12 @@
                                     {!! Form::submit('フォローする', ['class' => 'btn1 btn-default btn-sm']) !!}
                                 {!! Form::close() !!}
                             @endif
-                        @endif 
+                        @endif
+                    @else
+                        {!! Form::open(['route' => 'login', 'method' => 'get']) !!}
+                            {!! Form::submit('フォローする', ['class' => "btn btn-default btn-md"]) !!}
+                        {!! Form::close() !!}
+                    @endif
                        </dd>
                     </div>
                 </dl>
